chore: refactor sweep - Move all systems out of `main` - Implement plugin wrappers for all systems and startup - Created modules for each plugin and associated types - Created prelude module to load crate re-used pubs in one place
